Need to change the firebase config
The JS file to handle the login and sign up. Not sure how to get the firebase config Info from the firebase console after setting it up. I created a new test firebase project and used that info for the the firebase config. Still need to add the username to the profile. Apart from that, it should be functional.